added POST mapping for add the cacheData on RedisServer
