--- src/main/java/api/Result.java
@@ -1,33 +1,0 @@
-package api;
-
-import javax.validation.constraints.NotNull;
-
-public class Result {
-    public int getAccount() {
-        return account;
-    }
-
-    public void setAccount(int account) {
-        this.account = account;
-    }
-
-    public String getType() {
-        return type;
-    }
-
-    public void setType(String type) {
-        this.type = type;
-    }
-
-    public Long getValue() {
-        return value;
-    }
-
-    public void setValue(Long value) {
-        this.value = value;
-    }
-
-    @NotNull private int account;
-    private String type;
-    @NotNull private Long value;
-}
