+++ src/main/java/com/search/api/SearchController.java
@@ -1,4 +1,4 @@
-package api;
+package com.search.api;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.PostMapping;
@@ -7,12 +7,10 @@ import org.springframework.web.bind.annotation.RestController;
 
 import java.util.List;
 
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8;
-
 @RestController
 public class SearchController {
 
-    @Autowired private Service service;
+   private Service service;
 
     public static final String APPLICATION_JSON_VALUE = "application/json";
