+++ src/main/java/com/search/api/Service.java
@@ -1,6 +1,5 @@
-package api;
+package com.search.api;
 
-import javax.sql.rowset.serial.SerialException;
 import java.util.List;
 
 public interface Service {
